+++ src/test/java/mil/nga/geopackage/test/extension/elevation/ElevationTilesTiffTestUtils.java
@@ -179,8 +179,8 @@ public class ElevationTilesTiffTestUtils {
 					griddedCoverage.getTileMatrixSetName());
 			TestCase.assertEquals(GriddedCoverageDataType.FLOAT,
 					griddedCoverage.getDataType());
-			TestCase.assertNull(griddedCoverage.getScale());
-			TestCase.assertNull(griddedCoverage.getOffset());
+			TestCase.assertEquals(1.0, griddedCoverage.getScale());
+			TestCase.assertEquals(0.0, griddedCoverage.getOffset());
 			TestCase.assertTrue(griddedCoverage.getPrecision() >= 0);
 			griddedCoverage.getDataNull();
 
@@ -245,8 +245,8 @@ public class ElevationTilesTiffTestUtils {
 				griddedTile.getTableName());
 		long tableId = griddedTile.getTableId();
 		TestCase.assertTrue(tableId >= 0);
-		TestCase.assertNull(griddedTile.getScale());
-		TestCase.assertNull(griddedTile.getOffset());
+		TestCase.assertEquals(1.0, griddedTile.getScale());
+		TestCase.assertEquals(0.0, griddedTile.getOffset());
 		griddedTile.getMin();
 		griddedTile.getMax();
 		griddedTile.getMean();